+++ TeamProject.code-workspace
@@ -1,0 +1,8 @@
+{
+    "folders": [
+        {
+            "name": "contact-wizard",
+            "path": "."
+        }
+    ]
+}
