+++ models/Address.py
@@ -1,0 +1,5 @@
+from models.Field import Field
+
+class Address(Field):
+    def __init__(self, value):
+        super().__init__(value)
