+++ models/Phone.py
@@ -1,0 +1,5 @@
+from models.Field import Field
+
+class Phone(Field):
+    def __init__(self, value):
+        super().__init__(value)
