+++ models/contact.py
@@ -1,111 +1,81 @@
-from datetime import datetime
-from models.field import Field
-from collections import UserDict
+from models.Name import Name
+from models.Phone import Phone
+from models.Email import Email
+from models.Birthday import Birthday
+from models.Address import Address
 
-class Name(Field):
-    pass
-
-class Phone(Field):
-    def __init__(self, value):
-        # Placeholder for phone validation call
-        super().__init__(value)
-
-class Email(Field):
-    def __init__(self, value):
-        # Placeholder for email validation call
-        super().__init__(value)
-
-class Birthday(Field):
-    def __init__(self, value):
-        try:
-            datetime.strptime(value, '%d.%m.%Y')
-        except ValueError:
-            raise ValueError("Invalid birthday format. Please use DD.MM.YYYY.")
-        super().__init__(value)
-
-class Address(Field):
-    pass
 
 class Contact:
     def __init__(self, name):
         self.name = Name(name)
         self.address = None
         self.phones = []
-        self.emails = []
+        self.email = None
         self.birthday = None
-
+    
+    # Phone actions
     def add_phone(self, phone):
+        for phone_ in self.phones:
+            if phone_.value == phone:
+                print("Number already exists")
+                return
         self.phones.append(Phone(phone))
 
-    def remove_phone(self, phone):
-        if phone in self.phones:
-            self.phones.remove(phone)
+    def edit_phone(self, old_phone, new_phone):
+        pass
 
+    def remove_phone(self, phone):
+        pass
+        # if phone in self.phones:
+        #     self.phones.remove(phone)
+    
+    # Email actions
     def add_email(self, email):
-        self.emails.append(Email(email))
+        if self.email:
+            print("Email already exists")
+        else:
+            self.email = Email(email)
+        # self.emails.append(Email(email))
+    
+    def edit_email(self, email):
+        pass
 
     def remove_email(self, email):
-        if email in self.emails:
-            self.emails.remove(email)
+        self.email = None
+        # if email in self.emails:
+        #     self.emails.remove(email)
+    
+    # Address actions
+    def add_address(self, address):
+        if self.address:
+            print("Address already exists")
+        else:
+            self.address = Address(address)
+    
+    def edit_address(self, address):
+        pass
 
+    def remove_address(self, address):
+        self.address = None
+    
+    #Birthday actions
     def add_birthday(self, birthday):
-        if self.birthday is not None:
-            raise ValueError("Birthday already exists for this contact")
-        self.birthday = Birthday(birthday)
+        if self.birthday:
+            print("Birthday already exists")
+            # raise ValueError("Birthday already exists for this contact")
+        else:
+            self.birthday = Birthday(birthday)
+
+    def edit_birthday(self, birthday):
+        pass
+
+    def remove_birthday(self, birthday):
+        pass
 
     def __str__(self):
-        result = f"Contact name: {self.name.value}, phones: {'; '.join(str(p) for p in self.phones)}, emails: {'; '.join(str(e) for e in self.emails)}"
+        result = f"Contact name: {self.name.value}, phones: {'; '.join(p.value for p in self.phones)}, email: {self.email.value}"
         if self.address:
             result += f", address: {self.address.value}"
         if self.birthday:
-            result += f", birthday: {self.birthday}"
+            result += f", birthday: {self.birthday.value}"
         return result
-
-class AddressBook(UserDict):
-    def add_contact(self, contact):
-        if contact.name.value in self.data:
-            raise ValueError(f"Contact with name {contact.name.value} already exists in the address book")
-        self.data[contact.name.value] = contact
-
-    def find_contact(self, name):
-        return self.data.get(name)
-
-    def delete_contact(self, name):
-        # Placeholder delete_contact. Rewrite test code below
-        if name in self.data:
-            del self.data[name]
-
-    def search_contacts(self, query):
-        # Placeholder search_contacts. Rewrite test code below
-        return [contact for contact in self.data.values() if query.lower() in contact.name.value.lower()]
-
-    def edit_contact(self, old_name, new_contact):
-        # Placeholder edit_contact. Rewrite test code below
-        if old_name in self.data:
-            del self.data[old_name]
-            self.add_contact(new_contact)
-        else:
-            raise ValueError(f"Contact with name {old_name} not found in the address book")
-
-    def show_contacts(self):
-        if self.data:
-            result = "All contacts:\n"
-            for contact in self.data.values():
-                result += f"{contact}\n"
-            return result.strip()
-        else:
-            return "No contacts found."
-
-    def show_upcoming_birthdays(self, days):
-        # Placeholder > Showing birthday date. Rewrite test code below (test copied from hw)
-        today = datetime.today()
-        upcoming_birthdays = []
-
-        for contact in self.data.values():
-            if contact.birthday:
-                birthday_date = datetime.strptime(contact.birthday.value, '%d.%m.%Y')
-                days_until_birthday = (birthday_date - today).days
-                if 0 <= days_until_birthday <= days:
-                    upcoming_birthdays.append(contact)
-
-        return upcoming_birthdays
