+++ services/note_service.py
@@ -1,4 +1,5 @@
-from models.note import Note, NoteBook
+from models.Note import Note
+from models.NoteBook import NoteBook
 
 class NoteService:
     def __init__(self, note_book):
